move resource manager related files to resources
Also move localization files to lib/l10n folder
