+++ flutter/lib/benchmark/benchmark.dart
@@ -22,14 +22,14 @@ import 'package:mlcommons_ios_app/backend/bridge/isolate.dart';
 import 'package:mlcommons_ios_app/backend/list.dart';
 import 'package:mlcommons_ios_app/backend/run_settings.dart';
 import 'package:mlcommons_ios_app/benchmark/benchmark_result.dart';
-import 'package:mlcommons_ios_app/benchmark/resource_manager.dart';
+import 'package:mlcommons_ios_app/resources/resource_manager.dart';
 import 'package:mlcommons_ios_app/icons.dart';
 import 'package:mlcommons_ios_app/info.dart';
 import 'package:mlcommons_ios_app/protos/backend_setting.pb.dart' as pb;
 import 'package:mlcommons_ios_app/protos/mlperf_task.pb.dart' as pb;
 import 'package:mlcommons_ios_app/store.dart';
-import 'configurations_manager.dart';
-import 'utils.dart';
+import 'package:mlcommons_ios_app/resources/configurations_manager.dart';
+import '../resources/utils.dart';
 
 class Benchmark {
   final pb.BenchmarkSetting benchmarkSetting;
