+++ flutter/lib/resources/configurations_manager.dart
