+++ flutter/lib/resources/result_manager.dart
