+++ flutter/lib/ui/benchmarks_configuration_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:mlcommons_ios_app/benchmark/benchmark.dart';
-import 'package:mlcommons_ios_app/benchmark/configurations_manager.dart';
+import 'package:mlcommons_ios_app/resources/configurations_manager.dart';
 import 'package:mlcommons_ios_app/localizations/app_localizations.dart';
 import 'package:mlcommons_ios_app/store.dart';
 import 'package:mlcommons_ios_app/ui/error_dialog.dart';
